requirements.txt
faster-whisper==1.0.3
numpy>=1.24
sounddevice>=0.4.6
webrtcvad-wheels==2.0.14
requests>=2.31
pywin32>=306
psutil>=5.9.0
